Pass database and bootleg prepped files to prediction pipelines
