+++ components/predict.yaml
@@ -8,6 +8,8 @@ inputs:
   - {name: task_name, description: '', default: ''}
   - {name: model_name_or_path, description: '', default: ''}
   - {name: s3_input_datadir, description: '', default: ''}
+  - {name: s3_database_dir, description: '', default: ''}
+  - {name: s3_bootleg_prepped_data, description: '', default: ''}
   - {name: model_type, description: '', default: ''}
   - {name: dataset_subfolder, description: '', default: ''}
   - {name: val_batch_size, description: '', default: ''}
@@ -26,7 +28,7 @@ implementation:
     - -c 
     - |
       . /opt/genie-toolkit/lib.sh
-      parse_args "$0" "image owner eval_sets task_name model_name_or_path s3_input_datadir model_type dataset_subfolder metrics_output s3_metrics_output val_batch_size" "$@"
+      parse_args "$0" "image owner eval_sets task_name model_name_or_path s3_input_datadir s3_database_dir s3_bootleg_prepped_data model_type dataset_subfolder metrics_output s3_metrics_output val_batch_size" "$@"
       shift $n
       cd $HOME
 
@@ -50,6 +52,14 @@ implementation:
       esac
 
       aws s3 sync --no-progress ${s3_input_datadir}${dataset_subfolder} $dataset_dir --exclude "synthetic*.txt" --exclude "*bootleg*" --exclude "*chunked*"
+
+      if test "${s3_database_dir}" != None ; then
+        aws s3 sync --no-progress ${s3_database_dir}/eval/ ./database/
+      fi
+
+      if test "${s3_bootleg_prepped_data}" != None ; then
+        aws s3 sync --no-progress ${s3_bootleg_prepped_data} results_temp --exclude '*' --include '*bootleg_labels.jsonl'
+      fi
 
       # if not an s3 path assume pretrained model name and save it in genienlp format
       if [[ "${model_name_or_path}" != "s3://"* ]] ; then
@@ -99,6 +109,8 @@ implementation:
       --task_name, {inputValue: task_name},
       --model_name_or_path, {inputValue: model_name_or_path},
       --s3_input_datadir, {inputValue: s3_input_datadir},
+      --s3_database_dir, {inputValue: s3_input_datadir},
+      --s3_bootleg_prepped_data, {inputValue: s3_input_datadir},
       --model_type, {inputValue: model_type},
       --dataset_subfolder, {inputValue: dataset_subfolder},
       --s3_metrics_output, {outputPath: s3_metrics_output},
