+++ pipelines/training.py
@@ -267,96 +267,6 @@ def eval_step(
 
     return eval_op
 
-def prediction_step(
-    image,
-    owner,
-    eval_sets,
-    task_name,
-    model_name_or_path,
-    s3_input_datadir,
-    model_type,
-    dataset_subfolder,
-    val_batch_size,
-    genienlp_version,
-    additional_args
-):
-
-    predict_env = {
-        'GENIENLP_VERSION': genienlp_version,
-    }
-
-    predict_num_gpus=4
-    predict_op = components.load_component_from_file('components/predict.yaml')(
-            image=image,
-            owner=owner,
-            eval_sets=eval_sets,
-            task_name=task_name,
-            model_name_or_path=model_name_or_path,
-            s3_input_datadir=s3_input_datadir,
-            model_type=model_type,
-            dataset_subfolder=dataset_subfolder,
-            val_batch_size=val_batch_size,
-            additional_args=additional_args)
-    (predict_op.container
-        .set_memory_request('150G')
-        .set_memory_limit('150G')
-        .set_cpu_request('16')
-        .set_cpu_limit('16')
-        # not supported yet in the version of kfp we're using
-        #.set_ephemeral_storage_request('75G')
-        #.set_ephemeral_storage_limit('75G')
-        .set_gpu_limit(str(predict_num_gpus))
-    )
-    (add_env(add_ssh_volume(predict_op), predict_env)
-        .add_toleration(V1Toleration(key='nvidia.com/gpu', operator='Exists', effect='NoSchedule'))
-        .add_node_selector_constraint('beta.kubernetes.io/instance-type', 'g4dn.12xlarge'))
-
-    return predict_op
-
-
-
-def prediction_step_small(
-    image,
-    owner,
-    eval_sets,
-    task_name,
-    model_name_or_path,
-    s3_input_datadir,
-    model_type,
-    dataset_subfolder,
-    val_batch_size,
-    genienlp_version,
-    additional_args
-):
-
-    predict_env = {
-        'GENIENLP_VERSION': genienlp_version
-    }
-
-    predict_op = components.load_component_from_file('components/predict.yaml')(
-            image=image,
-            owner=owner,
-            eval_sets=eval_sets,
-            task_name=task_name,
-            model_name_or_path=model_name_or_path,
-            s3_input_datadir=s3_input_datadir,
-            model_type=model_type,
-            dataset_subfolder=dataset_subfolder,
-            val_batch_size=val_batch_size,
-            additional_args=additional_args)
-    (predict_op.container
-     .set_memory_limit('61G')
-     .set_memory_request('61G')
-     .set_cpu_limit('15')
-     .set_cpu_request('15')
-     )
-    (add_env(add_ssh_volume(predict_op), predict_env)
-     .add_toleration(V1Toleration(key='nvidia.com/gpu', operator='Exists', effect='NoSchedule'))
-     .add_node_selector_constraint('beta.kubernetes.io/instance-type', 'g4dn.4xlarge')
-     )
-
-    return predict_op
-
 
 
 def paraphrase_train_fewshot_step(
